Memcached changeset.
Includes a bit of refactoring to minimize the amount
of data per title we cache, since all titles are cached
under a single key. Even with a 1 MB limit per key,
this will hold a lot of titles. Caching is automatically
enabled if a MEMCACHED_SERVERS (or MEMCACHIER_SERVERS)
environment variable is set.

The list of IPs that voted for each title is stored in
a separate key, again to minimize the amount of data
placed in the titles data structure, but also to
limit writes to the titles key, so we aren't updating
the larger structure in Memcached for each vote.
Similarly, vote counts are stored separately as well.