Makes votes for invalid vote IDs count 5x as strongly towards flood detection.
Fix a mistake in a test (the invalid JSON wasn't being sent!).

Adds a bit more logging so we're aware of all messages and errors -- if
users are going to send something to bring it down, we should hope to
have at least some related log messages (even if still not very detailed).

Keeps track of which addresses have already been disconnected in the
current window, so that they can be more efficiently ignored (without
logging) for subsequent messages in the same window.